Remove garage.misc.ext
* Delete dead code
* Move `set_seed`/`get_seed` to `garage.experiment.deterministic`
* Move some utilities used exclusively by `garage.tf.optimizers` to
  `garage.tf.optimizers.utils`
