+++ garage/tf/optimizers/first_order_optimizer.py
@@ -4,10 +4,10 @@ import pyprind
 import tensorflow as tf
 
 from garage.core import Serializable
-from garage.misc import ext
 from garage.misc import logger
 from garage.optimizers import BatchDataset
 from garage.tf.misc import tensor_utils
+from garage.tf.optimizers.utils import LazyDict
 
 
 class FirstOrderOptimizer(Serializable):
@@ -84,7 +84,7 @@ class FirstOrderOptimizer(Serializable):
             if extra_inputs is None:
                 extra_inputs = list()
             self._input_vars = inputs + extra_inputs
-            self._opt_fun = ext.LazyDict(
+            self._opt_fun = LazyDict(
                 f_loss=lambda: tensor_utils.compile_function(
                     inputs + extra_inputs, loss), )
 
@@ -136,8 +136,8 @@ class FirstOrderOptimizer(Serializable):
                 elapsed = time.time() - start_time
                 callback_args = dict(
                     loss=new_loss,
-                    params=self._target.get_param_values(trainable=True)
-                    if self._target else None,
+                    params=self._target.get_param_values(
+                        trainable=True) if self._target else None,
                     itr=epoch,
                     elapsed=elapsed,
                 )
