+++ garage/tf/optimizers/penalty_lbfgs_optimizer.py
@@ -3,8 +3,9 @@ import scipy.optimize
 import tensorflow as tf
 
 from garage.core import Serializable
-from garage.misc import ext, logger
+from garage.misc import logger
 from garage.tf.misc import tensor_utils
+from garage.tf.optimizers.utils import LazyDict
 
 
 class PenaltyLbfgsOptimizer(Serializable):
@@ -81,7 +82,7 @@ class PenaltyLbfgsOptimizer(Serializable):
                         tf.cast(flat_grad, tf.float64),
                     ]
 
-            self._opt_fun = ext.LazyDict(
+            self._opt_fun = LazyDict(
                 f_loss=lambda: tensor_utils.compile_function(
                     inputs, loss, log_name="f_loss"),
                 f_constraint=lambda: tensor_utils.compile_function(
