+++ tests/benchmarks/test_benchmark_trpo.py
@@ -23,8 +23,8 @@ import pandas as pd
 import tensorflow as tf
 
 from garage.envs import normalize
+from garage.experiment import deterministic
 from garage.experiment import LocalRunner
-from garage.misc import ext
 from garage.misc import logger as garage_logger
 from garage.tf.algos import TRPO
 from garage.tf.baselines import GaussianMLPBaseline
@@ -101,7 +101,7 @@ def run_garage(env, seed, log_dir):
     :param log_dir: Log dir path.
     :return:import baselines.common.tf_util as U
     """
-    ext.set_seed(seed)
+    deterministic.set_seed(seed)
 
     with tf.Graph().as_default():
         with LocalRunner() as runner:
